api/requirements.txt
aniso8601==8.0.0
attrs==20.1.0
autopep8==1.5.4
certifi==2020.6.20
click==7.1.2
dnspython==2.0.0
elasticsearch==7.9.1
Flask==1.1.2
Flask-Cache==0.13.1
Flask-Caching==1.9.0
Flask-Elasticsearch==0.2.5
flask-restx==0.2.0
Flask-WTF==0.14.3
gunicorn==20.0.4
importlib-metadata==1.7.0
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
mongoengine==0.20.0
more-itertools==8.5.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pymongo==3.11.0
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==6.0.1
python-dotenv==0.14.0
python-memcached==1.59
pytz==2020.1
six==1.15.0
toml==0.10.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
zipp==3.1.0